+++ ribs/visualize/_grid_archive_heatmap.py
@@ -47,54 +47,58 @@ def grid_archive_heatmap(
     pass in ``pcm_kwargs={"edgecolor": "black", "linewidth": 0.1}``.
 
     Examples:
-        .. plot::
-            :context: close-figs
-
-            Heatmap of a 2D GridArchive
-
-            >>> import numpy as np
-            >>> import matplotlib.pyplot as plt
-            >>> from ribs.archives import GridArchive
-            >>> from ribs.visualize import grid_archive_heatmap
-            >>> # Populate the archive with the negative sphere function.
-            >>> archive = GridArchive(solution_dim=2,
-            ...                       dims=[20, 20],
-            ...                       ranges=[(-1, 1), (-1, 1)])
-            >>> x = np.random.uniform(-1, 1, 10000)
-            >>> y = np.random.uniform(-1, 1, 10000)
-            >>> archive.add(solution=np.stack((x, y), axis=1),
-            ...             objective=-(x**2 + y**2),
-            ...             measures=np.stack((x, y), axis=1))
-            >>> # Plot a heatmap of the archive.
-            >>> plt.figure(figsize=(8, 6))
-            >>> grid_archive_heatmap(archive)
-            >>> plt.title("Negative sphere function")
-            >>> plt.xlabel("x coords")
-            >>> plt.ylabel("y coords")
-            >>> plt.show()
+        Heatmap of a 2D GridArchive
 
         .. plot::
             :context: close-figs
 
-            Heatmap of a 1D GridArchive
+            import numpy as np
+            import matplotlib.pyplot as plt
+            from ribs.archives import GridArchive
+            from ribs.visualize import grid_archive_heatmap
 
-            >>> import numpy as np
-            >>> import matplotlib.pyplot as plt
-            >>> from ribs.archives import GridArchive
-            >>> from ribs.visualize import grid_archive_heatmap
-            >>> # Populate the archive with the negative sphere function.
-            >>> archive = GridArchive(solution_dim=2,
-            ...                       dims=[20], ranges=[(-1, 1)])
-            >>> x = np.random.uniform(-1, 1, 1000)
-            >>> archive.add(solution=np.stack((x, x), axis=1),
-            ...             objective=-x**2,
-            ...             measures=x[:, None])
-            >>> # Plot a heatmap of the archive.
-            >>> plt.figure(figsize=(8, 6))
-            >>> grid_archive_heatmap(archive)
-            >>> plt.title("Negative sphere function with 1D measures")
-            >>> plt.xlabel("x coords")
-            >>> plt.show()
+            # Populate the archive with the negative sphere function.
+            archive = GridArchive(solution_dim=2,
+                                  dims=[20, 20],
+                                  ranges=[(-1, 1), (-1, 1)])
+            x = np.random.uniform(-1, 1, 10000)
+            y = np.random.uniform(-1, 1, 10000)
+            archive.add(solution=np.stack((x, y), axis=1),
+                        objective=-(x**2 + y**2),
+                        measures=np.stack((x, y), axis=1))
+
+            # Plot a heatmap of the archive.
+            plt.figure(figsize=(8, 6))
+            grid_archive_heatmap(archive)
+            plt.title("Negative sphere function")
+            plt.xlabel("x coords")
+            plt.ylabel("y coords")
+            plt.show()
+
+        Heatmap of a 1D GridArchive
+
+        .. plot::
+            :context: close-figs
+
+            import numpy as np
+            import matplotlib.pyplot as plt
+            from ribs.archives import GridArchive
+            from ribs.visualize import grid_archive_heatmap
+
+            # Populate the archive with the negative sphere function.
+            archive = GridArchive(solution_dim=2,
+                                  dims=[20], ranges=[(-1, 1)])
+            x = np.random.uniform(-1, 1, 1000)
+            archive.add(solution=np.stack((x, x), axis=1),
+                        objective=-x**2,
+                        measures=x[:, None])
+
+            # Plot a heatmap of the archive.
+            plt.figure(figsize=(8, 6))
+            grid_archive_heatmap(archive)
+            plt.title("Negative sphere function with 1D measures")
+            plt.xlabel("x coords")
+            plt.show()
 
     Args:
         archive: A 1D or 2D :class:`~ribs.archives.GridArchive`.